backend/requirements.txt
Flask==1.1.1
psycopg2-binary==2.8.6
gunicorn==20.0.4
alembic
